[server] Removing monitoring EE
