+++ components/server/src/monitoring-endpoints.ts
@@ -6,11 +6,29 @@
 
 import * as express from "express";
 import { injectable } from "inversify";
+import { registerServerMetrics } from "../../src/prometheus-metrics";
+import * as prometheusClient from "prom-client";
+import { registerDBMetrics } from "@gitpod/gitpod-db/lib";
 
 @injectable()
 export class MonitoringEndpointsApp {
     public create(): express.Application {
-        // Monitoring is an EE feature
-        return express();
+        const registry = prometheusClient.register;
+
+        prometheusClient.collectDefaultMetrics({ register: registry });
+        registerDBMetrics(registry);
+        registerServerMetrics(registry);
+
+        const monApp = express();
+        monApp.get("/metrics", async (req, res) => {
+            try {
+                res.set("Content-Type", registry.contentType);
+                res.end(await registry.metrics());
+            } catch (ex) {
+                res.status(500).end(ex);
+            }
+        });
+
+        return monApp;
     }
 }
